requirements.txt
opencv-python==4.1.2.30
future==0.18.2
Pillow==8.0.1